Split visualization controller spec
