+++ Makefile
@@ -125,11 +125,12 @@ WORKING_SPECS_2 = \
 
 WORKING_SPECS_4 = \
   spec/rspec_configuration.rb \
+	spec/requests/carto/api/visualizations_controller_spec.rb \
+	spec/requests/carto/api/visualizations_controller_main_behaviour_spec.rb \
   services/wms/spec/unit/wms_spec.rb \
   services/sql-api/spec/sql_api_spec.rb \
   spec/requests/admin/visualizations_spec.rb \
   spec/requests/api/json/visualizations_controller_spec.rb \
-  spec/requests/carto/api/visualizations_controller_spec.rb \
   spec/requests/carto/api/tables_controller_spec.rb \
   spec/queries/carto/visualization_query_builder_spec.rb \
   spec/requests/admin/tables_spec.rb \
